Create a Lagged Variable

Function to create a lagged variable.

lagvar(x, obs, id, day, data, lag=1)

Arguments

x
vector to specify the variable to be lagged.

obs
numeric vector that indicates the observation number of the values specified via the x argument.

id
optional vector to specify a subject id variable.

day
optional vector to specify a day number variable.

data
optional data frame that contains the variables specified above.

lag
scalar specifying the lag (default is 1). Can also be a vector with multiple lag values that are permitted. See ‘Details’.

Details

The function can be used to create a lagged version of the variable specified via the x argument. The obs argument is used to specify a variable that indicates the observation number of the values specified via the x argument. There should not be repeated values of obs (within the same subject, in case the dataset includes multiple subjects).

The (optional) id argument is used to specify a subject id variable. If not specified, all observations are assumed to come from the same subject.

The (optional) day variable is used to specify a day number variable. If not specified, then values of the variable to be lagged are allowed to be lagged across days (in case there are multiple observation days). If the day variable is specified, then lagged values across different days are not allowed (i.e., are set to missing).

The lag argument is used to specify the number of lags (i.e., the lagged value of the variable at a particular observation number \(t\) is the value of the variable observed at observation number \(t - lag\)). If no lagged value is available at observation number \(t - lag\) (either because the dataset does not include that observation number or the value at that observation number is missing), then the lagged value will be missing.

One can also specify multiple values for the lag argument (usually a consecutive range, such as 1:3). In that case, the function will find the lagged value with the smallest lag and also indicate the lag of each lagged value. See ‘Examples’.

Value

If lag is a single number, the function returns a vector with the lagged values (that will typically be added back to the data frame as a new variable).

If multiple values are specified for lag, the function returns a data frame with elements xl for the lagged values and lag for the lag of those values.

Examples

# illustrative dataset (5 beeps for 2 days for 2 subjects)
# note: beeps not responded to are not included in this dataset
dat <- read.table(header=TRUE, text = "
subject dayno beep stress
1 1 1  3
1 1 2  2
1 1 3  4
1 1 4  1
1 1 5  2
1 2 6  5
1 2 7  2
1 2 9  3
1 2 10 3
2 1 1  2
2 1 2  3
2 1 3  2
2 2 6  1
2 2 9  5
2 2 10 2")
dat

#>    subject dayno beep stress
#> 1        1     1    1      3
#> 2        1     1    2      2
#> 3        1     1    3      4
#> 4        1     1    4      1
#> 5        1     1    5      2
#> 6        1     2    6      5
#> 7        1     2    7      2
#> 8        1     2    9      3
#> 9        1     2   10      3
#> 10       2     1    1      2
#> 11       2     1    2      3
#> 12       2     1    3      2
#> 13       2     2    6      1
#> 14       2     2    9      5
#> 15       2     2   10      2

# lag the stress variable
dat$stress_lag <- lagvar(stress, beep, subject, data=dat)
dat

#>    subject dayno beep stress stress_lag
#> 1        1     1    1      3         NA
#> 2        1     1    2      2          3
#> 3        1     1    3      4          2
#> 4        1     1    4      1          4
#> 5        1     1    5      2          1
#> 6        1     2    6      5          2
#> 7        1     2    7      2          5
#> 8        1     2    9      3         NA
#> 9        1     2   10      3          3
#> 10       2     1    1      2         NA
#> 11       2     1    2      3          2
#> 12       2     1    3      2          3
#> 13       2     2    6      1         NA
#> 14       2     2    9      5         NA
#> 15       2     2   10      2          5

# but this also allows lagged values across days
# this can be avoided by specifying the day variable
dat$stress_lag <- lagvar(stress, beep, subject, dayno, data=dat)
dat

#>    subject dayno beep stress stress_lag
#> 1        1     1    1      3         NA
#> 2        1     1    2      2          3
#> 3        1     1    3      4          2
#> 4        1     1    4      1          4
#> 5        1     1    5      2          1
#> 6        1     2    6      5         NA
#> 7        1     2    7      2          5
#> 8        1     2    9      3         NA
#> 9        1     2   10      3          3
#> 10       2     1    1      2         NA
#> 11       2     1    2      3          2
#> 12       2     1    3      2          3
#> 13       2     2    6      1         NA
#> 14       2     2    9      5         NA
#> 15       2     2   10      2          5

# lag the stress variable by two observations
dat$stress_lag <- lagvar(stress, beep, subject, dayno, data=dat, lag=2)
dat

#>    subject dayno beep stress stress_lag
#> 1        1     1    1      3         NA
#> 2        1     1    2      2         NA
#> 3        1     1    3      4          3
#> 4        1     1    4      1          2
#> 5        1     1    5      2          4
#> 6        1     2    6      5         NA
#> 7        1     2    7      2         NA
#> 8        1     2    9      3          2
#> 9        1     2   10      3         NA
#> 10       2     1    1      2         NA
#> 11       2     1    2      3         NA
#> 12       2     1    3      2          2
#> 13       2     2    6      1         NA
#> 14       2     2    9      5         NA
#> 15       2     2   10      2         NA

# lag the stress variable by up to 5 observations
# note: this creates two variables (xl and lag), which are both added to the dataset
dat$stress_lag <- lagvar(stress, beep, subject, dayno, data=dat, lag=1:5)$xl
dat$lags <- lagvar(stress, beep, subject, dayno, data=dat, lag=1:5)$lag
dat

#>    subject dayno beep stress stress_lag lags
#> 1        1     1    1      3         NA   NA
#> 2        1     1    2      2          3    1
#> 3        1     1    3      4          2    1
#> 4        1     1    4      1          4    1
#> 5        1     1    5      2          1    1
#> 6        1     2    6      5         NA   NA
#> 7        1     2    7      2          5    1
#> 8        1     2    9      3          2    2
#> 9        1     2   10      3          3    1
#> 10       2     1    1      2         NA   NA
#> 11       2     1    2      3          2    1
#> 12       2     1    3      2          3    1
#> 13       2     2    6      1         NA   NA
#> 14       2     2    9      5          1    3
#> 15       2     2   10      2          5    1

# illustrative dataset as above, but with missing beeps included
dat <- read.table(header=TRUE, text = "
subject dayno beep stress
1 1 1  3
1 1 2  2
1 1 3  4
1 1 4  1
1 1 5  2
1 2 6  5
1 2 7  2
1 2 8  NA
1 2 9  3
1 2 10 3
2 1 1  2
2 1 2  3
2 1 3  2
2 1 4  NA
2 1 5  NA
2 2 6  1
2 2 7  NA
2 2 8  NA
2 2 9  5
2 2 10 2")
dat

#>    subject dayno beep stress
#> 1        1     1    1      3
#> 2        1     1    2      2
#> 3        1     1    3      4
#> 4        1     1    4      1
#> 5        1     1    5      2
#> 6        1     2    6      5
#> 7        1     2    7      2
#> 8        1     2    8     NA
#> 9        1     2    9      3
#> 10       1     2   10      3
#> 11       2     1    1      2
#> 12       2     1    2      3
#> 13       2     1    3      2
#> 14       2     1    4     NA
#> 15       2     1    5     NA
#> 16       2     2    6      1
#> 17       2     2    7     NA
#> 18       2     2    8     NA
#> 19       2     2    9      5
#> 20       2     2   10      2

# lag the stress variable
dat$stress_lag <- lagvar(stress, beep, subject, dayno, data=dat)
dat

#>    subject dayno beep stress stress_lag
#> 1        1     1    1      3         NA
#> 2        1     1    2      2          3
#> 3        1     1    3      4          2
#> 4        1     1    4      1          4
#> 5        1     1    5      2          1
#> 6        1     2    6      5         NA
#> 7        1     2    7      2          5
#> 8        1     2    8     NA          2
#> 9        1     2    9      3         NA
#> 10       1     2   10      3          3
#> 11       2     1    1      2         NA
#> 12       2     1    2      3          2
#> 13       2     1    3      2          3
#> 14       2     1    4     NA          2
#> 15       2     1    5     NA         NA
#> 16       2     2    6      1         NA
#> 17       2     2    7     NA          1
#> 18       2     2    8     NA         NA
#> 19       2     2    9      5         NA
#> 20       2     2   10      2          5

# lag the stress variable by up to 5 observations
dat$stress_lag <- lagvar(stress, beep, subject, dayno, data=dat, lag=1:5)$xl
dat$lags <- lagvar(stress, beep, subject, dayno, data=dat, lag=1:5)$lag
dat

#>    subject dayno beep stress stress_lag lags
#> 1        1     1    1      3         NA   NA
#> 2        1     1    2      2          3    1
#> 3        1     1    3      4          2    1
#> 4        1     1    4      1          4    1
#> 5        1     1    5      2          1    1
#> 6        1     2    6      5         NA   NA
#> 7        1     2    7      2          5    1
#> 8        1     2    8     NA          2    1
#> 9        1     2    9      3          2    2
#> 10       1     2   10      3          3    1
#> 11       2     1    1      2         NA   NA
#> 12       2     1    2      3          2    1
#> 13       2     1    3      2          3    1
#> 14       2     1    4     NA          2    1
#> 15       2     1    5     NA          2    2
#> 16       2     2    6      1         NA   NA
#> 17       2     2    7     NA          1    1
#> 18       2     2    8     NA          1    2
#> 19       2     2    9      5          1    3
#> 20       2     2   10      2          5    1